userbot/requirements.txt
# Core dependencies
pyrogram==2.0.106
TgCrypto==1.2.5
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.1

# AI providers
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2

# Database and caching
asyncpg==0.29.0
redis==5.0.1

# Utilities
Pillow==10.1.0
moviepy==1.0.3
beautifulsoup4==4.12.2
lxml==4.9.3
humanize==4.9.0

# Logging
colorlog==6.8.0
python-json-logger==2.0.7

# Development
black==23.12.1
pylint==3.0.3
pytest==7.4.3
pytest-asyncio==0.21.1
